Update README.md
@alexfedoseev, let's consider removing the ordering one, unless we are going to implement our own linter rule.

Regarding naming, I think the Air Ugly, inconsistent one is OK, as is your. Consistency is best for this. I'd say we go with what you have on the naming and maybe modify the rule.

The bottom line is that if we're going to deviate, we need to do so with linter rules.